External api feedback (#52)
+ cleanup macros + old data structures
+ add new API to encode EOS nal
+ add ability to mute all Library printf [LIB_PRINTF_ENABLE]
+ add ability to print lib parameters in the log
+ remove math.h dependency
+ fix fps calculation that used to break rate control / fix print config format
+ unify CMake/.pc changes from gst/ffmpeg plugins
+ rename svt to svthevc in ffmpeg, SvtHevcEnc in pkg-config
+ fix lib install path on Windows
+ move lib's pkg-config and install down to the lib's Cmake config
+ remove -lm dependency
+ respect CMAKE_INSTALL_PREFIX and CMAKE_INSTALL_LIBDIR settings
+ cleanup macros + old data structures
+ add new API to encode EOS nal
+ add ability to mute all Library printf [LIB_PRINTF_ENABLE]
+ add ability to print lib parameters in the log
+ cleanup pragma from windows builds
+ remove math.h dependency
+ fix fps calculation that used to break rate control / fix print config format
+ unify CMake/.pc changes from gst/ffmpeg plugins
+ rename svt to svthevc in ffmpeg, SvtHevcEnc in pkg-config
+ fix lib install path on Windows
+ move lib's pkg-config and install down to the lib's Cmake config
+ remove -lm dependency
+ respect CMAKE_INSTALL_PREFIX and CMAKE_INSTALL_LIBDIR settings
+ add NON_AVX512_SUPPORT
+ move EB_ASM typedef inside library
+ fix buffer calculation based on lookahead distance
+ set codeEosNal to be 1 in simple application
+ set a default value for recon buffer to not crash at de-init when one is not defined
+ Change the EbH265EncStreamHeader API for the library to own the output buffer
+ Change the EbH265EncEosNal API for the library to own the output buffer
+ Change the EbH265GetPacket API for the library to own the output buffer
+ Added EbH265ReleaseOutBuffer API call to send feedback for the library to release the output buffer
+ Change EB_SLICE to EB_PICTURE

+++ Source/API/EbTime.h
@@ -2,16 +2,16 @@
 * Copyright(c) 2018 Intel Corporation
 * SPDX - License - Identifier: BSD - 2 - Clause - Patent
 */
-
+#include <stdint.h>
 #ifndef EbTime_h
 #define EbTime_h
 
-void EbStartTime(unsigned long long *Startseconds, unsigned long long *Startuseconds);
-void EbFinishTime(unsigned long long *Finishseconds, unsigned long long *Finishuseconds);
-void EbComputeOverallElapsedTime(unsigned long long Startseconds, unsigned long long Startuseconds,unsigned long long Finishseconds, unsigned long long Finishuseconds, double *duration);
-void EbComputeOverallElapsedTimeMs(unsigned long long Startseconds, unsigned long long Startuseconds, unsigned long long Finishseconds, unsigned long long Finishuseconds, double *duration);
-void EbSleep(unsigned long long milliSeconds);
-void EbInjector(unsigned long long processedFrameCount, unsigned int injectorFrameRate);
+void EbStartTime(uint64_t *Startseconds, uint64_t *Startuseconds);
+void EbFinishTime(uint64_t *Finishseconds, uint64_t *Finishuseconds);
+void EbComputeOverallElapsedTime(uint64_t Startseconds, uint64_t Startuseconds,uint64_t Finishseconds, uint64_t Finishuseconds, double *duration);
+void EbComputeOverallElapsedTimeMs(uint64_t Startseconds, uint64_t Startuseconds, uint64_t Finishseconds, uint64_t Finishuseconds, double *duration);
+void EbSleep(uint64_t milliSeconds);
+void EbInjector(uint64_t processedFrameCount, uint32_t injectorFrameRate);
 
 #endif // EbTime_h
 /* File EOF */
